packages/forge-config/pyproject.toml
[project]
name = "forge-config"
version = "0.1.0"
description = "Configuration schema, validation, and hot-reload for Forge AI"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
dependencies = [
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "watchdog>=4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "pytest-mock>=3.14",
    "ruff>=0.9",
    "mypy>=1.14",
    "types-PyYAML>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/forge_config"]
